+++ startup/gui/viewer.py
@@ -75,6 +75,7 @@ def __viewContextMenu( viewer, view, menuDefinition ) :
 
 	GafferSceneUI.LightUI.appendViewContextMenuItems( viewer, view, menuDefinition )
 	GafferSceneUI.SceneHistoryUI.appendViewContextMenuItems( viewer, view, menuDefinition )
+	GafferSceneUI.EditScopeUI.appendViewContextMenuItems( viewer, view, menuDefinition )
 
 GafferUI.Viewer.viewContextMenuSignal().connect( __viewContextMenu )